do not let Giant Bees make you berserker if you're exhausted

+++ monster.go
@@ -802,7 +802,7 @@ func (m *monster) HitSideEffects(g *game, ev event) {
 			g.Confusion(ev)
 		}
 	case MonsGiantBee:
-		if RandInt(5) == 0 && !g.Player.HasStatus(StatusBerserk) {
+		if RandInt(5) == 0 && !g.Player.HasStatus(StatusBerserk) && !g.Player.HasStatus(StatusExhausted) {
 			g.Player.Statuses[StatusBerserk]++
 			g.PushEvent(&simpleEvent{ERank: ev.Rank() + 25 + RandInt(40), EAction: BerserkEnd})
 			g.Print("You feel a sudden urge to kill things.")
